configure: C99 compatibility issues
As part of this effort:

https://fedoraproject.org/wiki/Toolchain/PortingToModernC

I've found an issue with one of the autoconf checks in the conntrack
package. It uses the exit functions without including stdlib. This is
deprecated in C99 because it no longer allows implicit function
declarations. Find attached a patch that changes the check to use return
instead of exit.

Closes: https://bugzilla.netfilter.org/show_bug.cgi?id=1654
Signed-off-by: Peter Fordham <peter.fordham@gmail.com>
Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>