conntrack: fix BPF for filtering IPv6 addresses
Each address in the filter is matched one 32-bit word at a time.  If any of the
first three words don't match, we jump to the end of the filter.  If the last
word does match, we jump to the end of the filter.  However, this is not right:
it means that if any of the first three words of an address don't match, all
subsequent addresses will be skipped.  Instead, jump to the next address.

Fix formatting of `nfct_bsf_cmp_k_stack`.

Link: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=690676
Closes: https://bugzilla.netfilter.org/show_bug.cgi?id=1062
Fixes: dd73e5708cc2 ("bsf: add support for IPv6 address filtering")
Signed-off-by: Jeremy Sowden <jeremy@azazel.net>
Signed-off-by: Florian Westphal <fw@strlen.de>